refactor(redis): Removed Redis implementations of things
