--- keel-redis/README.md
@@ -1,3 +1,0 @@
-# keel-redis
-
-Redis implementation of the resource repository.
