--- keel-redis/src/test/kotlin/com/netflix/spinnaker/keel/redis/RedisArtifactRepositoryTests.kt
@@ -1,29 +1,0 @@
-package com.netflix.spinnaker.keel.redis
-
-import com.netflix.spinnaker.keel.persistence.ArtifactRepositoryTests
-import com.netflix.spinnaker.keel.serialization.configuredObjectMapper
-import com.netflix.spinnaker.kork.jedis.EmbeddedRedis
-import com.netflix.spinnaker.kork.jedis.JedisClientDelegate
-import org.junit.jupiter.api.AfterAll
-import redis.clients.jedis.Jedis
-
-internal object RedisArtifactRepositoryTests : ArtifactRepositoryTests<RedisArtifactRepository>() {
-
-  override fun factory() = RedisArtifactRepository(
-    redisClient,
-    configuredObjectMapper()
-  )
-
-  private val embeddedRedis = EmbeddedRedis.embed()
-  private val redisClient = JedisClientDelegate(embeddedRedis.pool)
-
-  override fun flush() {
-    redisClient.withCommandsClient { redis -> (redis as Jedis).flushAll() }
-  }
-
-  @JvmStatic
-  @AfterAll
-  fun shutdown() {
-    embeddedRedis.destroy()
-  }
-}
