+++ keel-spring-test-support/keel-spring-test-support.gradle.kts
@@ -7,6 +7,6 @@ dependencies {
   api("org.springframework.boot:spring-boot-starter-test")
   api("org.springframework.security:spring-security-test")
   api("org.springframework.security:spring-security-config")
-  api("com.netflix.spinnaker.kork:kork-jedis")
+  api("com.netflix.eureka:eureka-client")
   api("com.ninja-squad:springmockk:1.1.0")
 }
