+++ keel-spring-test-support/src/main/kotlin/com/netflix/spinnaker/keel/spring/test/MockEurekaConfiguration.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.spinnaker.keel.redis.spring
+package com.netflix.spinnaker.keel.spring.test
 
 import com.netflix.appinfo.InstanceInfo
 import com.netflix.discovery.EurekaClient
